+++ R/req-perform.R
@@ -60,7 +60,7 @@
 #'     server doesn't exist), an error with class `"httr2_failure"`.
 #' @export
 #' @seealso [req_perform_parallel()] to perform multiple requests in parallel.
-#'   [req_perform_iterate()] to fetch all pages of a requests paginated via
+#'   [req_perform_iteratively()] to fetch all pages of a requests paginated via
 #'   [req_paginate()].
 #' @examples
 #' request("https://google.com") %>%